feat: implement Player class
